Refactored folder and file namse of PizzaSize component
