+++ src/modules/configurator/components/index.ts
@@ -1,7 +1,7 @@
 export * from './topping';
 export * from './toppings';
 export * from './discount';
-export * from './pizzaSize';
+export * from './pizzaSizeItem';
 export * from './pizzaPrice';
 export * from './clickCounter';
-export * from './pizzaSizes';
+export * from './pizzaSize';
