+++ src/modules/configurator/components/pizzaSize/PizzaSize.styles.ts
@@ -1,20 +1,12 @@
 import { css } from '@emotion/react';
 
-const pizza__size__btn = css({
-  padding: '14px 60px',
-  backgroundColor: '#ECECF3',
-  borderRadius: '36px',
-  border: 'none',
-  color: '#8C8C8C',
-});
-const pizza__size__btn__selected = css({
-  backgroundColor: '#FFFF',
-  boxShadow: '0px 8px 20px rgba(0, 0, 0, 0.1)',
+const pizza__size__selector = css({
+  display: 'flex',
+  justifyContent: 'space-between',
 });
 
 const styles = {
-  pizza__size__btn,
-  pizza__size__btn__selected,
+  pizza__size__selector,
 };
 
 export default styles;
