+++ src/modules/configurator/components/pizzaSize/PizzaSize.tsx
@@ -1,26 +1,15 @@
-/** @jsxImportSource @emotion/react */
-import { PizzaSizesType } from 'modules/configurator/types';
-import React, { useState } from 'react';
+import { pizzaSizesList } from 'modules/configurator/const/PizzaSizesList';
+import React from 'react';
+import { PizzaSizeItem } from '../pizzaSizeItem';
+
 import styles from './PizzaSize.styles';
-export const PizzaSize: React.FC<PizzaSizesType> = ({ size }) => {
-  const [selected, setSelected] = useState<string>('');
 
-  const setSelectedSize = (size: string) => {
-    size !== selected ? setSelected(size) : setSelected('');
-  };
-
+export const PizzaSize: React.FC = () => {
   return (
-    <div>
-      <button
-        css={
-          selected === size
-            ? [styles.pizza__size__btn, styles.pizza__size__btn__selected]
-            : styles.pizza__size__btn
-        }
-        onClick={() => setSelectedSize(size)}
-      >
-        {size}
-      </button>
+    <div css={styles.pizza__size__selector}>
+      {pizzaSizesList.map(({ size }) => {
+        return <PizzaSizeItem size={size} key={size} />;
+      })}
     </div>
   );
 };
