+++ src/modules/configurator/components/pizzaSizeItem/index.ts
@@ -1,0 +1,1 @@
+export * from './PizzaSizeItem';
