--- src/modules/configurator/components/pizzaSizes/PizzaSizes.styles.ts
@@ -1,17 +1,0 @@
-import { css } from '@emotion/react';
-
-const pizza__size__selector = css({
-  display: 'flex',
-  justifyContent: 'space-between',
-  gap: '16px',
-  backgroundColor: '#ECECF3',
-  padding: '8px 16px',
-  borderRadius: '72px',
-  minWidth: '438px',
-});
-
-const styles = {
-  pizza__size__selector,
-};
-
-export default styles;
