--- src/modules/configurator/components/pizzaSizes/index.ts
@@ -1,1 +1,0 @@
-export * from './PizzaSizes';
